feat: add auth service to client

+++ client/service/auth/TokenStorageService.tsx
@@ -1,45 +1,41 @@
-const TOKEN_KEY = "AuthToken";
-const USERNAME_KEY = "AuthUsername";
-const AUTHORITIES_KEY = "AuthAuthorities";
-const ROLE_KEY = 'Role'
+const ACCESSTOKEN_KEY = "AccessToken";
+const REFRESHTOKEN_KEY = "RefreshToken";
 
 export class TokenStorageService {
 
   constructor() {}
 
-  public signOut() {
-    window.localStorage.clear();
+  public clear() {
+    if (typeof window !== 'undefined') {
+      window.localStorage.clear();
+    }
   }
 
-  public saveToken(token: string) {
-    window.localStorage.removeItem(TOKEN_KEY);
-    window.localStorage.setItem(TOKEN_KEY, token);
+  public saveAccessToken(token: string) {
+    if (typeof window !== 'undefined') {
+      window.localStorage.removeItem(ACCESSTOKEN_KEY);
+      window.localStorage.setItem(ACCESSTOKEN_KEY, token);
+    }
   }
 
-  public getToken(): string {
-    return localStorage.getItem(TOKEN_KEY)!;
+  public saveRefreshToken(token: string) {
+    if (typeof window !== 'undefined') {
+      window.localStorage.removeItem(REFRESHTOKEN_KEY);
+      window.localStorage.setItem(REFRESHTOKEN_KEY, token);
+    }
   }
 
-  public saveLogin(login: string) {
-    window.localStorage.removeItem(USERNAME_KEY);
-    window.localStorage.setItem(USERNAME_KEY, login);
+  public getAccessToken() {
+    if (typeof window !== 'undefined') {
+      return window.localStorage.getItem(ACCESSTOKEN_KEY);
+    }
+    return null;
   }
 
-  public getLogin(): string {
-    return localStorage.getItem(USERNAME_KEY)!;
-  }
-
-  public saveRoles(roles: string) {
-    window.localStorage.removeItem(ROLE_KEY);
-    window.localStorage.setItem(ROLE_KEY, roles);
-  }
-
-  public getRoles(): string {
-    return localStorage.getItem(ROLE_KEY)!;
-  }
-
-  public saveAuthorities(authorities: string) {
-    window.localStorage.removeItem(AUTHORITIES_KEY);
-    window.localStorage.setItem(AUTHORITIES_KEY, JSON.stringify(arguments));
+  public getRefreshToken() {
+    if (typeof window !== 'undefined') {
+      return window.localStorage.getItem(REFRESHTOKEN_KEY);
+    }
+    return null;
   }
 }
